gs101: recovery: add wait for dwc3 device to probe
This patch adds a dependency on the dwc3 device probing before setting
the sys.usb.controller property. Without this, we have a race between
fastbootd starting and the dwc3 device probing. If fastbootd starts
before the dwc3 device probes, then the recovery init.rc will fail to
setup the USB gadget resulting in "fastboot devices" failing to detect
the device.

Bug: 202338035
Test: verify device is detected in fastbootd
Change-Id: I146451eae03f23dde915c00af3028681be41a8c1
Signed-off-by: Jason Edson <jaysonedson@gmail.com>